+++ sphinx/source/changelog.rst
@@ -15,6 +15,8 @@ Changelog
   defined by :ref:`label` directives)
 * Added the ``--erefs`` option to :ref:`sna2skool.py` (to always add comments
   that list entry point referrers)
+* Added the ``--package-dir`` option to :ref:`skool2asm.py` (to show the path
+  to the skoolkit package directory)
 * Removed support for the ``#NAME`` macro (what it did can be done by the
   :ref:`html` macro instead)
 * Removed the documentation sources and man page sources from the SkoolKit